switch from MLMediaLibrary to PhotoKit (with direct access to SQLite database)

+++ PhotosSync/ui/AppDelegate.swift
@@ -7,38 +7,65 @@
 //
 
 import Cocoa
+import Photos
 
 @NSApplicationMain
 class AppDelegate: NSObject, NSApplicationDelegate {
-
+    
     private let logger = Logger(loggerName: "AppDelegate", logLevel: .info)
-
+    
     func applicationDidFinishLaunching(_ aNotification: Notification) {
-        let preferences = PreferencesReader.readPreferencesFile()
-        logger.info("Read preferences file; content:\n\(preferences.toYaml())")
-
-        for plan in preferences.plans {
-            // separator for multiple export jobs
-            logger.info("")
-            logger.info("=====================================================================")
-            logger.info("")
-            
-            if plan.enabled {
-                logger.info("Start export using plan:\n\(plan.toYaml(indent: 10))")
-                do {
-                    let photosExporter = try PhotosExporterFactory.createPhotosExporter(plan: plan)
-                    photosExporter.exportPhotos()
-                } catch {
-                    print("Photos exporter could not be instantiated from preferences: \(String(describing: plan.name))")
-                }
-            } else {
-                logger.info("Ignore disabled plan: \(String(describing: plan.name))")
+        PHPhotoLibrary.requestAuthorization({(status: PHAuthorizationStatus) in
+            if (status != PHAuthorizationStatus.authorized) {
+                self.logger.warn("Not authorized to access the photo library. Abort.")
+                return;
             }
+            
+            self.executeAllPlans()
+            
+            DispatchQueue.main.sync {
+                NSApp.terminate(self)
+            }
+        });
+    }
+    
+    func executeAllPlans() {
+        var preferences: Preferences
+        do {
+            preferences = try PreferencesReader.readPreferencesFile()
+        } catch {
+            print("Error reading preferences: \(error)")
+            return;
         }
+        logger.info("Read preferences file; content:\n\(preferences.toYaml())")
+        
+        let photosMetadataReader = PhotosMetadataReader(config: preferences.config)
+        photosMetadataReader.readMetadata(completion: {(photosMetadata: PhotosMetadata) in
+            
+            //photosMetadata.rootCollection.printYaml(indent: 0)
+            
+            for plan in preferences.plans {
+                // separator for multiple export jobs
+                self.logger.info("")
+                self.logger.info("=====================================================================")
+                self.logger.info("")
+                
+                if plan.enabled {
+                    self.logger.info("Start export using plan:\n\(plan.toYaml(indent: 10))")
+                    do {
+                        let photosExporter = try PhotosExporterFactory.createPhotosExporter(plan: plan, photosMetadata: photosMetadata)
+                        photosExporter.exportPhotos()
+                    } catch {
+                        print("Error exporting photos for plan: \(String(describing: plan.name)); error: \(error)")
+                    }
+                } else {
+                    self.logger.info("Ignore disabled plan: \(String(describing: plan.name))")
+                }
+            }
+        });
+        
         
         //PreferencesReader.writePreferencesFile(preferences: preferences)
-
-        NSApp.terminate(self)
     }
     
     func applicationWillTerminate(_ aNotification: Notification) {
